# api/openapi.yaml
openapi: 3.1.0
info:
  title: NetScope API
  version: 0.1.0
  description: Concurrent network diagnostics API with persisted results and SSE progress.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
tags:
  - name: System
    description: Health, readiness, and runtime capabilities.
  - name: Runs
    description: Diagnostic run lifecycle, results, events, and exports.
security: []
paths:
  /healthz:
    get:
      tags: [System]
      operationId: getHealth
      summary: Check whether the API process is alive
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /readyz:
    get:
      tags: [System]
      operationId: getReadiness
      summary: Check PostgreSQL readiness
      responses:
        "200":
          description: API is ready
        "503":
          description: PostgreSQL is unavailable
  /api/v1/capabilities:
    get:
      tags: [System]
      operationId: getCapabilities
      summary: Get available diagnostic checks and runtime limits
      responses:
        "200":
          description: Capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
  /api/v1/runs:
    post:
      tags: [Runs]
      operationId: createRun
      summary: Queue a diagnostic run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Run queued
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Target blocked by network policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Run queue or database unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [Runs]
      operationId: listRuns
      summary: List persisted diagnostic runs
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/RunStatus"
      responses:
        "200":
          description: Run history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunPage"
  /api/v1/runs/{id}:
    parameters:
      - $ref: "#/components/parameters/RunID"
    get:
      tags: [Runs]
      operationId: getRun
      summary: Get a diagnostic run and its results
      responses:
        "200":
          description: Diagnostic run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runs/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/RunID"
    post:
      tags: [Runs]
      operationId: cancelRun
      summary: Cancel a queued or active run
      responses:
        "204":
          description: Cancellation accepted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Run has already finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/runs/{id}/events:
    parameters:
      - $ref: "#/components/parameters/RunID"
    get:
      tags: [Runs]
      operationId: streamRunEvents
      summary: Stream live run progress
      description: |
        Server-Sent Events stream. Event names are `run.started`,
        `check.started`, `check.completed`, `check.failed`, `run.completed`,
        and `run.cancelled`. REST and PostgreSQL remain the source of truth.
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runs/{id}/export:
    parameters:
      - $ref: "#/components/parameters/RunID"
      - in: query
        name: format
        schema:
          type: string
          const: json
          default: json
    get:
      tags: [Runs]
      operationId: exportRun
      summary: Export a run as JSON
      responses:
        "200":
          description: JSON export
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    RunID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid target, options, or request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Diagnostic run not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Health:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
          const: netscope
        version:
          type: string
    Capability:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
        reason:
          type: string
    Capabilities:
      type: object
      required: [version, checks, runtime]
      properties:
        version:
          type: string
        checks:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Capability"
        runtime:
          type: object
          required:
            [
              defaultTimeoutMs,
              maxTimeoutMs,
              runWorkers,
              probeConcurrency,
              networkPolicy,
            ]
          properties:
            defaultTimeoutMs:
              type: integer
            maxTimeoutMs:
              type: integer
            runWorkers:
              type: integer
            probeConcurrency:
              type: integer
            networkPolicy:
              type: string
              enum: [local, public]
    CheckType:
      type: string
      enum: [dns, tcp, http, tls]
    RunStatus:
      type: string
      enum:
        [queued, running, completed, partial, failed, cancelled, interrupted]
    CheckStatus:
      type: string
      enum: [pending, running, passed, warning, failed, skipped, cancelled]
    RunOptions:
      type: object
      required:
        [timeoutMs, httpMethod, followRedirects, maxRedirects, ipVersion]
      properties:
        timeoutMs:
          type: integer
          minimum: 500
          maximum: 30000
          default: 5000
        tcpPorts:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: integer
            minimum: 1
            maximum: 65535
        httpMethod:
          type: string
          enum: [GET, HEAD]
          default: GET
        followRedirects:
          type: boolean
          default: true
        maxRedirects:
          type: integer
          minimum: 0
          maximum: 10
          default: 5
        ipVersion:
          type: string
          enum: [auto, ipv4, ipv6]
          default: auto
    CreateRunRequest:
      type: object
      required: [target, checks]
      properties:
        target:
          type: string
          minLength: 1
          maxLength: 2048
          examples: [example.com, "https://example.com", 1.1.1.1]
        checks:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/CheckType"
        options:
          $ref: "#/components/schemas/RunOptions"
    CreateRunResponse:
      type: object
      required: [id, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          const: queued
        createdAt:
          type: string
          format: date-time
    CheckResult:
      type: object
      required:
        [id, runId, type, status, durationMs, data, startedAt, completedAt]
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/CheckType"
        status:
          $ref: "#/components/schemas/CheckStatus"
        durationMs:
          type: integer
          format: int64
        summary:
          type: string
        data:
          oneOf:
            - $ref: "#/components/schemas/DNSResult"
            - $ref: "#/components/schemas/TCPResult"
            - $ref: "#/components/schemas/HTTPResult"
            - $ref: "#/components/schemas/TLSResult"
            - type: object
              additionalProperties: true
        errorCode:
          type: string
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    DNSResult:
      type: object
      required: [a, aaaa, mx, ns, txt, ptr]
      properties:
        a:
          type: array
          items:
            type: string
            format: ipv4
        aaaa:
          type: array
          items:
            type: string
            format: ipv6
        cname:
          type: string
        mx:
          type: array
          items:
            type: object
            required: [host, preference]
            properties:
              host:
                type: string
              preference:
                type: integer
        ns:
          type: array
          items:
            type: string
        txt:
          type: array
          items:
            type: string
        ptr:
          type: array
          items:
            type: string
        errors:
          type: object
          additionalProperties:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
              message:
                type: string
    TCPResult:
      type: object
      required: [ports]
      properties:
        ports:
          type: array
          items:
            type: object
            required: [port, status, connectTimeMs]
            properties:
              port:
                type: integer
                minimum: 1
                maximum: 65535
              status:
                type: string
                enum: [passed, failed, cancelled]
              resolvedIp:
                type: string
              connectTimeMs:
                type: integer
                format: int64
              errorCode:
                type: string
              errorMessage:
                type: string
    HTTPResult:
      type: object
      required: [requestedUrl, method, redirectChain, timings, bodyTruncated]
      properties:
        requestedUrl:
          type: string
          format: uri
        finalUrl:
          type: string
          format: uri
        method:
          type: string
        statusCode:
          type: integer
        protocol:
          type: string
        contentType:
          type: string
        contentLength:
          type: integer
          format: int64
        redirectChain:
          type: array
          items:
            type: object
            required: [url, statusCode]
            properties:
              url:
                type: string
                format: uri
              statusCode:
                type: integer
        resolvedIp:
          type: string
        remoteAddress:
          type: string
        timings:
          type: object
          required: [dnsMs, connectMs, tlsMs, ttfbMs, totalMs]
          properties:
            dnsMs:
              type: integer
            connectMs:
              type: integer
            tlsMs:
              type: integer
            ttfbMs:
              type: integer
            totalMs:
              type: integer
        bodySha256:
          type: string
        bodyPreview:
          type: string
        bodyTruncated:
          type: boolean
    TLSResult:
      type: object
      required:
        [
          serverName,
          sans,
          daysRemaining,
          hostnameValid,
          chainValid,
          selfSigned,
          expired,
          warnings,
        ]
      properties:
        tlsVersion:
          type: string
        cipherSuite:
          type: string
        serverName:
          type: string
        resolvedIp:
          type: string
        subject:
          type: string
        issuer:
          type: string
        serialNumber:
          type: string
        sans:
          type: array
          items:
            type: string
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        daysRemaining:
          type: integer
        hostnameValid:
          type: boolean
        chainValid:
          type: boolean
        selfSigned:
          type: boolean
        expired:
          type: boolean
        warnings:
          type: array
          items:
            type: string
    DiagnosticRun:
      type: object
      required:
        [
          id,
          target,
          normalizedHost,
          status,
          checks,
          options,
          results,
          createdAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        target:
          type: string
        normalizedHost:
          type: string
        normalizedUrl:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/RunStatus"
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckType"
        options:
          $ref: "#/components/schemas/RunOptions"
        results:
          type: array
          items:
            $ref: "#/components/schemas/CheckResult"
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
    RunPage:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticRun"
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              examples: [invalid_target, run_not_found, target_blocked]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string